--- src/main/java/com/eventscheduler/Model.java
@@ -1,4 +1,0 @@
-package com.eventscheduler;
-
-public interface Model {
-}
