--- src/main/java/com/eventscheduler/Observable.java
@@ -1,7 +1,0 @@
-package com.eventscheduler;
-
-public interface Observable {
-    void addObserver(Observer  observer);
-    void removeObserver(Observer  observer);
-    void notifyObservers(Model model);
-}
