--- src/main/java/com/eventscheduler/Observer.java
@@ -1,5 +1,0 @@
-package com.eventscheduler;
-
-public interface Observer {
-    void update(Model model);
-}
