+++ src/main/java/com/eventscheduler/controller/EventObservable.java
@@ -1,9 +1,11 @@
-package com.eventscheduler;
+package com.eventscheduler.controller;
+
+import com.eventscheduler.model.Model;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class CalendarActivityObservable implements Observable {
+public class EventObservable implements Observable {
     private List<Observer> observers = new ArrayList<>();
 
     @Override
