+++ src/main/java/com/eventscheduler/controller/Observable.java
@@ -1,0 +1,9 @@
+package com.eventscheduler.controller;
+
+import com.eventscheduler.model.Model;
+
+public interface Observable {
+    void addObserver(Observer observer);
+    void removeObserver(Observer  observer);
+    void notifyObservers(Model model);
+}
