+++ src/main/java/com/eventscheduler/controller/Observer.java
@@ -1,0 +1,7 @@
+package com.eventscheduler.controller;
+
+import com.eventscheduler.model.Model;
+
+public interface Observer {
+    void update(Model model);
+}
