+++ src/main/java/com/eventscheduler/model/EventModel.java
@@ -1,6 +1,5 @@
 package com.eventscheduler.model;
 
-import com.eventscheduler.Model;
 import org.bson.codecs.pojo.annotations.BsonId;
 import org.bson.types.ObjectId;
