+++ src/main/java/com/eventscheduler/model/Model.java
@@ -1,0 +1,4 @@
+package com.eventscheduler.model;
+
+public interface Model {
+}
